[Refactor] 画像の読込、タイル関連定義をgraphics-win.cpp/hに統合
