+++ src/Makefile.am
@@ -933,8 +933,7 @@ EXTRA_hengband_SOURCES = \
 	main-win/main-win-sound.cpp main-win/main-win-sound.h \
 	main-win/main-win-tokenizer.cpp main-win/main-win-tokenizer.h \
 	main-win/main-win-windows.h \
-	main-win/read-graphics.cpp main-win/read-graphics.h \
-	main-win/tile-info.h \
+	main-win/graphics-win.cpp main-win/graphics-win.h \
 	wall.bmp \
 	stdafx.cpp stdafx.h
